+++ packages/labs/rollup.config.js
@@ -15,6 +15,6 @@
 import {litProdConfig} from '../../rollup-common.js';
 
 export default litProdConfig({
-  entryPoints: ['index'],
+  entryPoints: ['index', 'callbacks/callbacks'],
   external: ['updating-element', 'lit-html', 'lit-element'],
 });