+++ packages/labs/src/index.ts
@@ -11,3 +11,4 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
+export * from './callbacks/callbacks.js';